requirements.txt
langchain
faiss-cpu
openai
accelerate==0.20.3
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.4
einops==0.6.1
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
psutil==5.9.5
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing-extensions==4.6.3
urllib3==2.0.3
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.23
click==8.1.3
clickhouse-connect==0.6.4
cmake==3.26.4
colorclass==2.2.2
commonmark==0.9.1
compressed-rtf==1.0.6
cryptography==41.0.1
dataclasses-json==0.5.8
Deprecated==1.2.14
duckdb==0.8.1
easygui==0.98.3
ebcdic==1.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
extract-msg==0.41.1
faiss-cpu==1.7.4
fastapi==0.98.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gpt4all==0.3.4
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.15.1
idna==3.4
IMAPClient==2.3.1
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.197
langchainplus-sdk==0.0.17
lark-parser==0.12.0
lit==16.0.6
llama-cpp-python==0.1.50
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.0.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
oletools==0.60.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pandoc==2.3
pcodedmp==1.2.6
pdfminer.six==20221105
Pillow==9.5.0
plumbum==1.8.2
ply==3.11
posthog==3.0.1
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
PyMuPDF==1.22.3
pypandoc==1.11
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
red-black-tree-mod==1.20
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
RTFDE==0.0.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.11.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.2
unstructured==0.6.6
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.14.1
XlsxWriter==3.1.2
yarl==1.9.2
zstandard==0.21.0
